+++ config-sample.py
@@ -15,6 +15,9 @@ WEEREG_MYSQL_PASSWORD = 'weewx'
 # It should be slightly less than client's post_interval.
 WEEREG_MIN_DELAY = 3600 * 23
 
+# How long to wait for the screen capture subprocess
+SCREEN_CAPTURE_TIMEOUT = 180
+
 # Configuration for method GET /api/v2/stations
 WEEREG_STATIONS_MAX_AGE = 3600 * 24 * 30  # = One month
 WEEREG_STATIONS_LIMIT = 2000